Integrate with latest Parallaxer.rx.
