+++ .ruby-version
@@ -1,0 +1,1 @@
+2.6.3
