--- Sources/Views/InfinitePageKeyView.swift
@@ -1,407 +1,0 @@
-import UIKit
-
-extension Int {
-
-    var isEven: Bool {
-        return self % 2 == 0
-    }
-}
-
-extension CGRect {
-
-    /// The rect's center point.
-    var center: CGPoint {
-        return CGPoint(x: midX, y: midY)
-    }
-}
-
-let randomColors = [
-    UIColor.red,
-    UIColor.gray,
-    UIColor.green,
-    UIColor.blue,
-    UIColor.purple,
-    UIColor.yellow,
-    UIColor.magenta,
-    UIColor.brown
-]
-
-final class CircleView: UIView {
-
-    /// The circle's scale transform.
-    var scaleTransform: CGAffineTransform = .identity {
-        didSet { updateTransform() }
-    }
-
-    /// The circle's translation transform.
-    var translationTransform: CGAffineTransform = .identity {
-        didSet { updateTransform() }
-    }
-
-    private func updateTransform() {
-        transform = scaleTransform.concatenating(translationTransform)
-    }
-}
-
-//------------------------------------------------------------------------------------------------------------
-// Number of pages: 1
-// Number of positions: 1 (min/max)
-// 0
-//(a)
-// a
-// X                    <- Left edge interval:  [nil]
-// X                    <- Wrapping interval:   [nil]
-// X                    <- Right edge interval: [nil]
-
-//------------------------------------------------------------------------------------------------------------
-// Number of pages: 2
-// Number of positions: 2 (min/max)
-// 0 1
-//(a)b
-// a(b)
-// [ ]                  <- Left edge interval:  [0,1]
-//   X                  <- Wrapping interval:   [nil]
-//   X                  <- Right edge interval: [nil]
-
-//------------------------------------------------------------------------------------------------------------
-// Number of pages: 3
-// Number of positions: 3 (min/max)
-// 0 1 2
-//(a)b c
-// a(b)c
-// a b(c)
-// [ ]                  <- Left edge interval:  [0,1]
-//   X                  <- Wrapping interval:   [nil]
-//   [ ]                <- Right edge interval: [1,2]
-
-//------------------------------------------------------------------------------------------------------------
-// Number of pages: 4
-// Number of positions: 3 (min)
-// 0 1 2 3
-//(a)b c
-// a(b)c
-//   b(c)a
-// [ ]                  <- Left edge interval:  [0,1]
-//   [ ]                <- Wrapping interval:   [1,2]
-//     [ ]              <- Right edge interval: [2,3]
-
-//------------------------------------------------------------------------------------------------------------
-// Number of pages: 5
-// Number of positions: 3 (min)
-// 0 1 2 3 4
-//(a)b c
-// a(b)c
-// . b(c)a
-// . . c(a)b
-//     c a(b)
-// [ ]                  <- Left edge interval:  [0,1]
-//   [   ]              <- Wrapping interval:   [1,3]
-//       [ ]            <- Right edge interval: [3,4]
-
-//------------------------------------------------------------------------------------------------------------
-// Number of pages: 5
-// Number of positions: 4
-// 0 1 2 3 4
-//(a)b c d
-// a(b)c d
-// a b(c)d
-// . b c(d)a
-//   b c d(a)
-// [   ]                <- Left edge interval:  [0,2]
-//     [ ]              <- Wrapping interval:   [2,3]
-//       [ ]            <- Right edge interval: [3,4]
-
-//------------------------------------------------------------------------------------------------------------
-// Number of pages: 5
-// Number of positions: 5 (max)
-// 0 1 2 3 4
-//(a)b c d e
-// a(b)c d e
-// a b(c)d e
-// a b c(d)e
-// a b c d(e)
-// [   ]                <- Left edge interval:  [0,2]
-//     X                <- Wrapping interval:   [nil]
-//     [   ]            <- Right edge interval: [2,4]
-
-//------------------------------------------------------------------------------------------------------------
-// Number of pages: 10
-// Number of positions: 3 (minimum)
-// 0 1 2 3 4 5 6 7 8 9
-//(a)b c
-// a(b)c
-// . b(c)a
-// . . c(a)b
-// . . . a(b)c
-// . . . . b(c)a
-// . . . . . c(a)b
-// . . . . . . a(b)c
-// . . . . . . . b(c)a
-//               b c(a)
-// [ ]                  <- Left edge interval:  [0,1]
-//   [             ]    <- Wrapping interval:   [1,8]
-//                 [ ]  <- Right edge interval: [8,9]
-
-//------------------------------------------------------------------------------------------------------------
-// Number of pages: 10
-// Number of positions: 5
-// 0 1 2 3 4 5 6 7 8 9
-//(a)b c d e
-// a(b)c d e
-// a b(c)d e
-// . b c(d)e a
-// . . c d(e)a b
-// . . . d e(a)b c
-// . . . . e a(b)c d
-// . . . . . a b(c)d e
-//           a b c(d)e
-//           a b c d(e)
-// [   ]                <- Left edge interval:  [0,2]
-//     [         ]      <- Wrapping interval:   [2,7]
-//               [   ]  <- Right edge interval: [7,9]
-
-//------------------------------------------------------------------------------------------------------------
-// Number of pages: 10
-// Number of positions: 9
-// 0 1 2 3 4 5 6 7 8 9
-//(a)b c d e f g h i
-// a(b)c d e f g h i
-// a b(c)d e f g h i
-// a b c(d)e f g h i
-// a b c d(e)f g h i
-// . b c d e(f)g h i a
-// . b c d e f(g)h i a
-// . b c d e f g(h)i a
-// . b c d e f g h(i)a
-// . b c d e f g h i(a)
-// [       ]            <- Left edge interval:  [0,4]
-//         [ ]          <- Wrapping interval:   [4,5]
-//           [       ]  <- Right edge interval: [5,9]
-
-// Settable Properties
-//------------------------------------------------------------------------------------------------------------
-// Number of pages: 10
-// Number of positions: 10 (maximum)
-// 0 1 2 3 4 5 6 7 8 9
-//(a)b c d e f g h i j
-// a(b)c d e f g h i j
-// a b(c)d e f g h i j
-// a b c(d)e f g h i j
-// a b c d(e)f g h i j
-// a b c d e(f)g h i j
-// a b c d e f(g)h i j
-// a b c d e f g(h)i j
-// a b c d e f g h(i)j
-// a b c d e f g h i(j)
-// [       ]            <- Left edge interval:  [0,4]
-//         X            <- Wrapping interval:   [nil]
-//         [         ]  <- Right edge interval: [4,9]
-
-//------------------------------------------------------------------------------------------------------------
-//  Infinite animation. Each line below shows a state change that occurs as the user scrolls.
-//     (0, 1, 2, 3, 4, 5, 6, 7, 8, 9)   - all elements (numberOfPages: 10)
-//     [0, 1, 2, 3, 4]...               - visible elements (numberOfCircles: 5)
-//      ^                               - cursor position
-//         ^
-//            ^
-//    .[1, 2, 3, 4, 5]...
-//   ..[2, 3, 4, 5, 6]...
-//  ...[3, 4, 5, 6, 7]..
-//  ...[4, 5, 6, 7, 8].
-//  ...[5, 6, 7, 8, 9]
-//               ^
-//                  ^
-//------------------------------------------------------------------------------------------------------------
-//  Travel ratio: at the edges of the list
-//     (0, 1, 2, 3, 4, 5, 6, 7, 8, 9)   - all elements
-//     {0, 1, 2}            {7, 8, 9}   - edge elements where cursor position changes.
-//------------------------------------------------------------------------------------------------------------
-// Wrapping behavior:
-//     ...|0 1 2 3 4|...                - visible indices (numberOfCircles: 5)
-//       ...|1 2 3 4 0|...
-//         ...|2 3 4 0 1|...
-//           ...|3 4 0 1 2|...
-//             ...|4 0 1 2 3|...
-//               ...|0 1 2 3 4|...
-//
-//  TravelWidth = numberOfCircles * CircleWidth
-//  NumWraps(i) = (Travel / circleWidth) * (numberOfCircles - i)
-//  MidX(i) = (Travel * circleWidth) % numberOfCircles
-final class InfinitePageKeyView: UIView {
-    
-    /// Set the number of pages this key view is capable of representing.
-    @IBInspectable var numberOfPages: Int = 3 {
-        didSet { buildUI() }
-    }
-
-    /// The number of circles displayed in the view.
-    var numberOfCircles: Int {
-        return min(maxNumberOfCircles, numberOfPages)
-    }
-
-    @IBInspectable var maxNumberOfCircles: Int = 7 {
-        didSet { buildUI() }
-    }
-
-    var stepDistance: CGFloat {
-//        return bounds.width / CGFloat(numberOfCircles)
-        return floor(bounds.width / CGFloat(numberOfCircles))
-    }
-
-    /// Distance the container has traveled.
-    var travel: CGFloat = 0 {
-        didSet {
-            let translation3d = CATransform3DMakeTranslation(-travel, 0, 0)
-            travelView?.layer.transform = translation3d
-        }
-    }
-
-    var numberOfTravelSteps: Int {
-        return max(numberOfPages - numberOfCircles, 0)
-    }
-
-    var travelRatio: Double {
-        return 1 / floor(Double(numberOfCircles) / 2)
-    }
-
-    /// The center cursor location.
-    var centerPosition: CGFloat {
-//        let circleRadius = round(stepDistance / 2)
-//        let centerPosition = bounds.midX - circleRadius
-        let centerPosition = round(leftPosition + ((CGFloat(numberOfCircles) / 2) * stepDistance))
-        print("centerPosition: \(centerPosition)")
-        return centerPosition
-    }
-    
-    /// The left-most cursor position.
-    var leftPosition: CGFloat {
-//        let circleRadius = round(stepDistance / 2)
-        let leftPosition = bounds.minX
-        print("leftPosition: \(leftPosition)")
-        return leftPosition
-    }
-    
-    /// The right-most cursor position.
-    var rightPosition: CGFloat {
-//        let circleRadius = floor(stepDistance / 2)
-//        let rightPosition = (circleContainerView?.bounds.maxX ?? 0)
-//        let rightPosition = leftPosition + ceil(CGFloat(numberOfCircles - 1) * stepDistance)
-        let rightPosition = bounds.maxX - (stepDistance)
-        print("rightPosition: \(rightPosition)")
-        return rightPosition
-//        return leftPosition + ceil(CGFloat(numberOfCircles - 1) * stepDistance)
-    }
-
-    /// The current cursor position.
-    var cursorPosition: CGFloat = 0.0 {
-        didSet { updateCursor() }
-    }
-    
-    /// The current cursor size.
-    var cursorScale: CGFloat = 1 {
-        didSet { updateCursor() }
-    }
-
-    /// All circle views, excluding the cursor.
-    private(set) var circleViews: [CircleView] = []
-
-    private weak var cursorView: CircleView?
-    private weak var travelView: UIView?
-    private weak var circleContainerView: UIView?
-
-    override func awakeFromNib() {
-        super.awakeFromNib()
-        buildUI()
-    }
-    
-    override func layoutSubviews() {
-        super.layoutSubviews()
-
-        let viewCenter = bounds.center
-
-        travelView?.bounds = bounds
-        travelView?.center = viewCenter
-
-        circleContainerView?.bounds = bounds
-        circleContainerView?.center = viewCenter
-        
-        let circleDiameter = stepDistance
-        let circleRadius = round(circleDiameter / 2)
-
-        let circleSize = CGSize(width: circleDiameter, height: circleDiameter)
-        let circleCenter = CGPoint(x: circleRadius, y: viewCenter.y)
-        circleContainerView?.subviews.enumerated().forEach { i, circleView in
-            circleView.bounds.size = circleSize
-            circleView.center = circleCenter
-            circleView.bounds = circleView.bounds.integral
-            circleView.layer.cornerRadius = circleRadius
-        }
-        
-        cursorView?.bounds.size = circleSize
-        cursorView?.center = circleCenter
-        cursorView?.layer.cornerRadius = circleRadius
-    }
-
-    /// Calculate the squared distance between the specified subview and the cursor. This is the squared
-    /// distance, which is useful for (performant) relative distance comparisons.
-    ///
-    /// - Parameter subview: A subview of the page key view.
-    /// - Returns: The quared distance between the specified subview and the cursor.
-    func distanceSquaredFromCursor(for subview: UIView) -> CGFloat {
-        guard let cursorView = self.cursorView else {
-            return 0
-        }
-
-        let cursorBoundsRelativeToView = cursorView.convert(cursorView.bounds, to: self)
-        let subviewBoundsRelativeToView = subview.convert(subview.bounds, to: self)
-
-        let cursorCenter = CGPoint(x: cursorBoundsRelativeToView.midX, y: cursorBoundsRelativeToView.midY)
-        let subviewCenter = CGPoint(x: subviewBoundsRelativeToView.midX, y: subviewBoundsRelativeToView.midY)
-
-        // Distance formula, without the square-root, performant way to compare relative distances.
-        return pow(cursorCenter.x - subviewCenter.x, 2) + pow(cursorCenter.y - subviewCenter.y, 2)
-    }
-    
-    private func buildUI() {
-        self.cursorView?.removeFromSuperview()
-        self.travelView?.removeFromSuperview()
-        
-        let travelView = UIView()
-        addSubview(travelView)
-
-        let circleContainerView = UIView()
-        travelView.addSubview(circleContainerView)
-
-        circleViews = (0 ..< numberOfCircles).map { _ in
-            let circleView = createCircleView(isHollow: true)
-            circleContainerView.addSubview(circleView)
-            circleView.backgroundColor = randomColors.randomElement()
-            return circleView
-        }
-
-        let cursorView = createCircleView(isHollow: false)
-        addSubview(cursorView)
-
-        self.travelView = travelView
-        self.circleContainerView = circleContainerView
-        self.cursorView = cursorView
-
-        setNeedsLayout()
-        layoutIfNeeded()
-
-        self.backgroundColor = UIColor(red: 0, green: 0, blue: 1, alpha: 0.5)
-        travelView.backgroundColor = UIColor(red: 0, green: 1, blue: 0, alpha: 0.5)
-    }
-    
-    private func createCircleView(isHollow: Bool) -> CircleView {
-        let view = CircleView()
-        view.backgroundColor = isHollow ? UIColor.init(white: 1, alpha: 0.5) : UIColor.white
-        return view
-    }
-
-    private func updateCursor() {
-        let scale = CGAffineTransform(scaleX: cursorScale, y: cursorScale)
-        let translation = CGAffineTransform(translationX: cursorPosition, y: 0)
-        cursorView?.transform = scale.concatenating(translation)
-    }
-}
